Changes for PR #1 "Refactor application"
