+++ cmd/going-live/main.go
@@ -6,7 +6,7 @@ import (
 )
 
 var (
-	port = "3000"
+	port = "3001"
 )
 
 func main() {
