+++ cmd/going-live/routes.go
@@ -3,7 +3,7 @@ package main
 import (
 	"github.com/go-chi/chi/v5"
 
-	"github.com/advance467/Going-Live/cmd/going-live/titles"
+	"github.com/advance467/Going-Live/cmd/going-live/features/titles"
 )
 
 func MainRoutes() *chi.Mux {
